Revert "LPS-123758 Move classes that are not exported to internal package - portal-cache"
This reverts commit 100dfd000db586ce75cba4d46e850fe5fdc273ed.
