+++ modules/apps/portal-cache/portal-cache-multiple/src/main/java/com/liferay/portal/cache/multiple/configuration/PortalCacheClusterConfiguration.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.multiple.internal.configuration;
+package com.liferay.portal.cache.multiple.configuration;
 
 import aQute.bnd.annotation.metatype.Meta;
