+++ modules/apps/portal-cache/portal-cache-multiple/src/main/java/com/liferay/portal/cache/multiple/internal/cluster/link/ClusterLinkPortalCacheClusterChannelFactory.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.cache.multiple.internal.cluster.link;
 
+import com.liferay.portal.cache.multiple.configuration.PortalCacheClusterConfiguration;
 import com.liferay.portal.cache.multiple.internal.PortalCacheClusterException;
-import com.liferay.portal.cache.multiple.internal.configuration.PortalCacheClusterConfiguration;
 import com.liferay.portal.cache.multiple.internal.constants.PortalCacheDestinationNames;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.cluster.ClusterLink;
